% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai.R
\name{aicopy}
\alias{aicopy}
\title{Copy Error with Information to Clipboard}
\usage{
aicopy()
}
\description{
Call \code{aicopy()} to copy your error message and other information useful to LLMs to clipboard.
This could be pasted into ChatGPT or Claude (for example) in order to help you fix the error.
}
