% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai.R
\name{ai}
\alias{ai}
\title{Get AI Help after an Error}
\usage{
ai()
}
\value{
string
}
\description{
Call \code{ai()} to help you fix an error. This will stream LLM response to your console intended to help
understand and fix the error. Will default to your most recently installed model if the option \code{"errbud_model"}
is not set.
}
